+++ CHANGELOG.md
@@ -557,3 +557,4 @@ Developmental progress by date is stored in this file.
 
 ## 2026-03-21
 - Added an app logo that handles Android / iOS using the flutter_launcher_icons package, and Web by manually changing the favicon image
+- Added the hidePassword variable to allow showing and hiding input password text